made a image preloader class for the canvas - no useEffect required, linting for canvas.tsx is good now
